+++ tests/Unit/Expression/LexerTest.php
@@ -3,7 +3,6 @@
 namespace PhpBench\Tests\Unit\Expression;
 
 use PhpBench\Expression\Lexer;
-use PhpBench\Expression\Token;
 use PhpBench\Expression\Tokens;
 use PHPUnit\Framework\TestCase;
 
@@ -18,22 +17,6 @@ class LexerTest extends TestCase
 EOT
         );
         self::assertCount(3, $tokens);
-    }
-
-    public function testLexParameter(): void
-    {
-        $tokens = $this->lex(<<<'EOT'
-foo.asd.asd
-EOT
-        );
-        self::assertEquals('foo.asd.asd', $tokens->chomp(Token::T_PARAMETER)->value);
-
-        $tokens = $this->lex(<<<'EOT'
-foobar(foo.asd.asd)
-EOT
-        );
-        $tokens->chomp(Token::T_FUNCTION);
-        self::assertEquals('foo.asd.asd', $tokens->chomp(Token::T_PARAMETER)->value);
     }
 
     private function lex(string $string): Tokens